Refactor main.ts to extract specific functionalities
